st190x: refactor code generation to seperate class
